+++ NEWS.md
@@ -5,6 +5,14 @@ New features
 
 * Allow arbitrary baseline hazard functions in `node_cox()`. This also makes it possible to perform discrete-event simulations using `sim_discrete_event()` with variables that have continuous time-dependent baseline hazards, by using the `model` argument.
 
+Enhancements
+
+* Added the `all_levels` argument to the `rcategorical()` function to allow users to keep all levels, even when some never occur in the generated data.
+
+Bug Fixes
+
+* Fixed a bug that occurred when using `dag_from_data()` with categorical parents in nodes based on generalized linear models and negative binomial regression models.
+
 # simDAG 0.5.2
 
 Bug Fixes